Refactor buildah bud for vendoring
Projects like libpod would benefit greatly if we could vendor in buildah
rather than execing it.  This refactor moves some of the key parsing
functions that buds uses to pkg/parse so that the project vendoring
buildah can also benefit from them.
